requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
aiosqlite
pydantic
python-dotenv
typer[all]
rich
pandas
pandas-ta
setuptools
numpy<2.0 # Constrain numpy version for pandas-ta compatibility
